input manager major refactor
