+++ cvvideoplayer/__init__.py
@@ -3,4 +3,4 @@ from .recorder import SimpleRecorder, AbstractRecorder
 from .video_player import VideoPlayer
 from .utils.bbox_utils import Bbox
 from .utils.video_player_utils import KeyFunction
-from .input_manager import InputManager
+from .input_management import InputParser, InputHandler
