+++ cvvideoplayer/video_player.py
@@ -7,7 +7,7 @@ from functools import partial
 import cv2
 import numpy as np
 
-from .input_manager import InputManager, SupportedOS
+from .input_management import InputParser, SupportedOS, InputHandler
 from .frame_editors import BaseFrameEditCallback
 from .frame_reader import FrameReader
 from .recorder import AbstractRecorder
@@ -41,6 +41,7 @@ class VideoPlayer:
         It can also be an instance of AbstractRecorder for custom recording functionality.
         """
         self.frame_reader = get_frame_reader(video_source)
+        self.input_handler = InputHandler()
         self._recorder = get_recorder(record)
         self._last_frame = len(self.frame_reader) - 1
         self._current_frame_num = start_from_frame - 1
@@ -67,26 +68,26 @@ class VideoPlayer:
             self._recorder.teardown()
         for callback in self._frame_edit_callbacks:
             callback.teardown()
-        InputManager().stop()
+        InputParser().stop()
 
     def run(self) -> None:
         try:
             self._setup()
-            InputManager().start()
+            InputParser().start()
             self._run_player_loop()
         finally:
             self.__exit__()
 
     def _run_player_loop(self):
         while not self._exit:
-            key = InputManager().get_input()
+            single_input = InputParser().get_input()
             if is_window_closed_by_mouse_click(window_name=self._window_name):
                 cv2.waitKey(5)
                 break
             elif get_in_focus_window_id(self._current_system) != self._window_id:
                 continue
             else:
-                InputManager().handle_key_str(key)
+                self.input_handler.handle_input(single_input)
                 self._play_continuously() if self._play else self._show_current_frame()
 
     def _setup_callbacks(self):
@@ -99,7 +100,7 @@ class VideoPlayer:
                     f"{callback.__class__.__name__} is trying to register a key function"
                     f" (key = {key_function.key}) which is not an instance of KeyFunction"
                 )
-                InputManager().register_key_function(key_function, callback.__class__.__name__)
+                self.input_handler.register_key_function(key_function, callback.__class__.__name__)
 
             callback.setup(self, self._current_frame)
 
@@ -137,7 +138,7 @@ class VideoPlayer:
         cv2.waitKey(1)  # for some reason Windows OS requires an additional waitKey to work properly
 
     def _play_continuously(self) -> None:
-        while (not InputManager().has_input()) and self._play:
+        while (not InputParser().has_input()) and self._play and not self._exit:
             self._change_current_frame(change_by=self._play_speed)
             self._show_current_frame()
 
@@ -210,4 +211,4 @@ class VideoPlayer:
         ]
 
         for key_function in default_key_functions:
-            InputManager().register_key_function(key_function, "Video Control")
+            self.input_handler.register_key_function(key_function, "Video Control")
